+++ ReleaseNotes.md
@@ -2,6 +2,11 @@
 
 # Release portal-disease
 
-## Release 1.3.0
+## Release 1.3.2
+- Updated ospo-resources for adding additional notes and disclaimer
 
+## Release 1.3.1
+- Add new font and background color
+
+## Release 1.3.0
 - First official GitHub-Release